Update deploy script to check if any local changes on branch

+++ scripts/gh-pages-deploy.js
@@ -1,21 +1,25 @@
 import {execa} from 'execa';
 (async () => {
-    try {
-      await execa("git", ["checkout", "--orphan", "gh-pages"]);
-      console.log("Building...");
-      await execa("npm", ["run", "build"]);
-      // Understand if it's dist or build folder
-    //   const folderName = fs.existsSync("dist") ? "dist" : "build";
-      await execa("git", ["--work-tree", 'dist', "add", "--all"]);
-      await execa("git", ["--work-tree", 'dist', "commit", "-m", "gh-pages"]);
-      console.log("Pushing to gh-pages...");
-      await execa("git", ["push", "origin", "HEAD:gh-pages", "--force"]);
-    //   await execa("rm", ["-r", folderName]);
-      await execa("git", ["checkout", "-f", "main"]);
-      await execa("git", ["branch", "-D", "gh-pages"]);
-      console.log("Successfully deployed");
-    } catch (e) {
-      console.log(e.message);
-      process.exit(1);
-    }
+	try { 
+		const { stdout } = await execa("git", ["diff-index", "HEAD"]);
+		if (stdout) {
+			console.log("Please stash or commit changes first!");
+			process.exit(1);
+		} else { 
+			await execa("git", ["checkout", "--orphan", "gh-pages"]);
+			console.log("Building...");
+			await execa("npm", ["run", "build"]);
+			await execa("git", ["--work-tree", 'dist', "add", "--all"]);
+			await execa("git", ["--work-tree", 'dist', "commit", "-m", "gh-pages"]);
+			console.log("Pushing to gh-pages...");
+			await execa("git", ["push", "origin", "HEAD:gh-pages", "--force"]);
+			await execa("git", ["checkout", "-f", "main"]);
+			await execa("git", ["branch", "-D", "gh-pages"]);
+			console.log("Successfully deployed");
+		}
+	}  catch (e) {
+		console.log(e.message);
+		process.exit(1);
+	}
+
   })();
